feat!(backend): move internal endpoints into internal controller service
